Bug Fix - Refresh routes
Using require to import the JSON database for the redirs causes Node JS to import it once and never do so again. This made it so the application had to be refreshed in order to refresh to new routes. I've fixed this by reading in the json with fs every time.